Delete a buuuuunch of legacy code
