+++ identity/webapp/src/frontend/MyAccount/Loading.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+export const Loading: React.FC = () => (
+  <label>
+    Loading…
+    <progress></progress>
+  </label>
+);
