+++ identity/webapp/src/frontend/Registration/AccountCreated.tsx
@@ -1,9 +1,9 @@
 import Icon from '@weco/common/views/components/Icon/Icon';
 import SpacingComponent from '@weco/common/views/components/SpacingComponent/SpacingComponent';
 import React from 'react';
-import { PageWrapper } from '../Shared/PageWrapper';
+import { PageWrapper } from '../components/PageWrapper';
 import { SuccessMessage } from './Registration.style';
-import { Container, Wrapper } from '../Shared/Layout.style';
+import { Container, Wrapper } from '../components/Layout.style';
 
 export const AccountCreated: React.FC = () => {
   return (
