+++ identity/webapp/src/frontend/Registration/Registration.style.ts
@@ -1,39 +1,9 @@
-import { SolidButton } from '@weco/common/views/components/ButtonSolid/ButtonSolid';
 import CheckboxRadio from '@weco/common/views/components/CheckboxRadio/CheckboxRadio';
-import { FieldError } from 'react-hook-form';
 import styled, { css } from 'styled-components';
 import { Link } from 'react-router-dom';
 
-export const FieldMargin = styled.div`
-  margin-bottom: 1em;
-`;
-
-export const Label = styled.label.attrs({ className: 'font-hnl fonts-loaded font-size-4' })`
-  display: block;
-  font-weight: bold;
-`;
-
-export const TextInput = styled.input<{ invalid?: FieldError }>`
-  width: 100%;
-  height: 55px;
-  margin: 0.333em 0;
-  padding: 0.7em;
-  border: ${props => (props.invalid ? 'solid 2px #d1192c' : 'solid 1px #8f8f8f')};
-  border-radius: 6px;
-`;
-
 export const ExternalLink = styled.a`
   white-space: nowrap;
-`;
-
-export const Button = styled(SolidButton)`
-  width: 100%;
-  justify-content: center;
-`;
-
-export const InvalidFieldAlert = styled.span.attrs({ role: 'alert', className: 'font-hnl' })`
-  color: #d1192c;
-  font-weight: bold;
 `;
 
 const AlertBox = styled.div.attrs({ role: 'alert', className: 'font-hnl' })`
