+++ identity/webapp/src/frontend/Registration/Registration.tsx
@@ -3,28 +3,24 @@ import { useForm, Controller } from 'react-hook-form';
 import { ErrorMessage } from '@hookform/error-message';
 import { Link, useHistory } from 'react-router-dom';
 import { AccountCreated } from './AccountCreated';
-import { PageWrapper } from '../Shared/PageWrapper';
+import { PageWrapper } from '../components/PageWrapper';
 import {
-  Button,
   Cancel,
   Checkbox,
   ErrorAlert,
-  FieldMargin,
-  InvalidFieldAlert,
-  Label,
   ExternalLink,
   PasswordRulesList,
-  TextInput,
   CheckboxLabel,
   InProgress,
 } from './Registration.style';
-
-import { Container, Title, Wrapper } from '../Shared/Layout.style';
+import { FieldMargin, Label, TextInput, InvalidFieldAlert, Button } from '../components/Form.style';
+import { Container, Title, Wrapper } from '../components/Layout.style';
 import SpacingComponent from '@weco/common/views/components/SpacingComponent/SpacingComponent';
 import Icon from '@weco/common/views/components/Icon/Icon';
 import { useRegisterUser, RegistrationError } from './useRegisterUser';
-import { PasswordInput } from './PasswordInput';
+import { PasswordInput } from '../components/PasswordInput';
 import { usePageTitle } from '../hooks/usePageTitle';
+import { validEmailPattern, validPasswordPattern } from '../components/ValidationPatterns';
 
 type RegistrationInputs = {
   firstName: string;
@@ -33,9 +29,6 @@ type RegistrationInputs = {
   password: string;
   termsAndConditions: boolean;
 };
-
-const validEmailPattern = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-const validPasswordPattern = /(?=.{8,})(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9]).*/;
 
 export function Registration(): JSX.Element {
   const { register, control, handleSubmit, formState, setError } = useForm<RegistrationInputs>({
