--- identity/webapp/src/frontend/Registration/RegistrationSummaryParagraph.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export const RegistrationSummaryParagraph: React.FC = () => (
-  // Placeholder text for now, Wellcome to provide full text here.
-  <p className="font-hnl font-size-4">
-    Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aliquam at porttitor sem. Aliquam erat volutpat. Donec
-    placerat nisl magna, et faucibus arcu condimentum sed.
-  </p>
-);
