--- identity/webapp/src/frontend/Shared/ErrorMessage.tsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-export const ErrorMessage = styled.div.attrs({ role: 'alert' })`
-  width: 100%;
-  padding: 12px 20px;
-  border: 1px solid #f5c6cb;
-  border-radius: 4px;
-  background-color: #f8d7da;
-  color: #721c24;
-  justify-content: center;
-`;
