--- identity/webapp/src/frontend/Shared/LogoContainer.tsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-
-export const LogoContainer = styled.div`
-   {
-    margin: auto;
-    padding: 42px;
-    width: 200px;
-  }
-`;
