--- identity/webapp/src/frontend/Shared/SuccessMessage.tsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-export const SuccessMessage = styled.div.attrs({ role: 'alert' })`
-  width: 100%;
-  padding: 12px 20px;
-  border: 1px solid #ffeeba;
-  border-radius: 4px;
-  background-color: #fff3cd;
-  color: #856404;
-  justify-content: center;
-`;
