--- identity/webapp/src/frontend/Shared/WarningMessage.tsx
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-export const WarningMessage = styled.div`
-   {
-    width: 100%;
-    padding: 12px 20px;
-    border: 1px solid #f5c6cb;
-    border-radius: 4px;
-    background-color: #fff3cd;
-    color: #856404;
-    justify-content: center;
-  }
-`;
