+++ identity/webapp/src/frontend/index.tsx
@@ -1,11 +1,8 @@
 import React from 'react';
 import { render } from 'react-dom';
 import { Registration } from './Registration/Registration';
-import { OldRegistration } from './Registration/OldRegistration';
 import { AccountValidated } from './Registration/AccountValidated';
-import { AccountManagement } from './AccountManagement/AccountManagement';
-import { ErrorPage } from './Shared/ErrorPage';
-
+import { ErrorPage } from './components/ErrorPage';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import { AppContextProvider } from '@weco/common/views/components/AppContext/AppContext';
 import { ThemeProvider, createGlobalStyle } from 'styled-components';
@@ -33,11 +30,9 @@ if (root) {
         <BrowserRouter basename={prefix || ''}>
           <Switch>
             <Route exact path="/register" component={Registration} />
-            <Route exact path="/register/old" component={OldRegistration} />
             <Route exact path="/validated" component={AccountValidated} />
             <Route exact path="/error" component={ErrorPage} />
             <Route exact path="/" component={MyAccount} />
-            <Route exact path="/old" component={AccountManagement} />
           </Switch>
         </BrowserRouter>
       </AppContextProvider>
